internal/dag: rename dag.DAG to dag.Builder
Updates #445

Rename dag.DAG to dag.Builder to accurately reflect what it does. This
also free's up the public symbol dag.DAG to now define the output of
dag.Builder.Compute.

Split the package into two files, builder.go and dag.go with their
respective types and helpers.
