+++ internal/contour/listener_test.go
@@ -431,7 +431,7 @@ func TestListenerVisit(t *testing.T) {
 			}
 			v := listenerVisitor{
 				ListenerCache: lc,
-				visitable:     d.Compute(),
+				Visitable:     d.Compute(),
 			}
 			got := v.Visit()
 			if !reflect.DeepEqual(tc.want, got) {
