+++ internal/dag/dag.go
@@ -16,16 +16,9 @@
 package dag
 
 import (
-	"fmt"
-	"strconv"
-	"strings"
-	"sync"
 	"time"
 
 	"k8s.io/api/core/v1"
-	"k8s.io/api/extensions/v1beta1"
-	"k8s.io/apimachinery/pkg/util/intstr"
-	"k8s.io/client-go/tools/cache"
 
 	"github.com/envoyproxy/go-control-plane/envoy/api/v2/auth"
 	ingressroutev1 "github.com/heptio/contour/apis/contour/v1beta1"
@@ -35,22 +28,6 @@ import (
 // between Kubernetes Ingress objects, the backend Services, and Secret objects.
 // The DAG models these relationships as Roots and Vertices.
 type DAG struct {
-	// IngressRouteRootNamespaces specifies the namespaces where root
-	// IngressRoutes can be defined. If empty, roots can be defined in any
-	// namespace.
-	IngressRouteRootNamespaces []string
-
-	mu sync.Mutex
-
-	ingresses     map[meta]*v1beta1.Ingress
-	ingressroutes map[meta]*ingressroutev1.IngressRoute
-	secrets       map[meta]*v1.Secret
-	services      map[meta]*v1.Service
-}
-
-// dag holds the results of recomputing a materialised directed acyclic graph
-// from the state stored in a DAG.
-type dag struct {
 	// roots are the roots of this dag
 	roots []Vertex
 
@@ -58,475 +35,17 @@ type dag struct {
 	statuses []Status
 }
 
-func (d *dag) Visit(fn func(Vertex)) {
+// Visit calls fn on each root of this DAG.
+func (d *DAG) Visit(fn func(Vertex)) {
 	for _, r := range d.roots {
 		fn(r)
 	}
 }
 
-func (d *dag) Statuses() []Status {
+// Statuses returns a slice of Status objects associated with
+// the computation of this DAG.
+func (d *DAG) Statuses() []Status {
 	return d.statuses
-}
-
-// meta holds the name and namespace of a Kubernetes object.
-type meta struct {
-	name, namespace string
-}
-
-// Insert inserts obj into the DAG. If an object with a matching type, name, and
-// namespace exists, it will be overwritten.
-func (d *DAG) Insert(obj interface{}) {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-	switch obj := obj.(type) {
-	case *v1.Secret:
-		m := meta{name: obj.Name, namespace: obj.Namespace}
-		if d.secrets == nil {
-			d.secrets = make(map[meta]*v1.Secret)
-		}
-		d.secrets[m] = obj
-	case *v1.Service:
-		m := meta{name: obj.Name, namespace: obj.Namespace}
-		if d.services == nil {
-			d.services = make(map[meta]*v1.Service)
-		}
-		d.services[m] = obj
-	case *v1beta1.Ingress:
-		m := meta{name: obj.Name, namespace: obj.Namespace}
-		if d.ingresses == nil {
-			d.ingresses = make(map[meta]*v1beta1.Ingress)
-		}
-		d.ingresses[m] = obj
-	case *ingressroutev1.IngressRoute:
-		m := meta{name: obj.Name, namespace: obj.Namespace}
-		if d.ingressroutes == nil {
-			d.ingressroutes = make(map[meta]*ingressroutev1.IngressRoute)
-		}
-		d.ingressroutes[m] = obj
-	default:
-		// not an interesting object
-	}
-}
-
-// Remove removes obj from the DAG. If no object with a matching type, name, and
-// namespace exists in the DAG, no action is taken.
-func (d *DAG) Remove(obj interface{}) {
-	switch obj := obj.(type) {
-	default:
-		d.remove(obj)
-	case cache.DeletedFinalStateUnknown:
-		d.Remove(obj.Obj) // recurse into ourselves with the tombstoned value
-	}
-}
-
-func (d *DAG) remove(obj interface{}) {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-	switch obj := obj.(type) {
-	case *v1.Secret:
-		m := meta{name: obj.Name, namespace: obj.Namespace}
-		delete(d.secrets, m)
-	case *v1.Service:
-		m := meta{name: obj.Name, namespace: obj.Namespace}
-		delete(d.services, m)
-	case *v1beta1.Ingress:
-		m := meta{name: obj.Name, namespace: obj.Namespace}
-		delete(d.ingresses, m)
-	case *ingressroutev1.IngressRoute:
-		m := meta{name: obj.Name, namespace: obj.Namespace}
-		delete(d.ingressroutes, m)
-	default:
-		// not interesting
-	}
-}
-
-// Compute returns an object with responds to Visit and Statuses.
-func (d *DAG) Compute() interface {
-	Visit(func(Vertex))
-	Statuses() []Status
-} {
-	d.mu.Lock()
-	current := d.compute()
-	d.mu.Unlock()
-	return current
-}
-
-// serviceMap memoise access to a service map, built
-// as needed from the list of services cached
-// from k8s.
-type serviceMap struct {
-	// backing services from k8s api.
-	services map[meta]*v1.Service
-
-	// cached Services.
-	_services map[portmeta]*Service
-}
-
-// lookup returns a Service that matches the meta and port supplied.
-// If no matching Service is found lookup returns nil.
-func (sm *serviceMap) lookup(m meta, port intstr.IntOrString) *Service {
-	if port.Type == intstr.Int {
-		if s, ok := sm._services[portmeta{name: m.name, namespace: m.namespace, port: int32(port.IntValue())}]; ok {
-			return s
-		}
-	}
-	svc, ok := sm.services[m]
-	if !ok {
-		return nil
-	}
-	for i := range svc.Spec.Ports {
-		p := &svc.Spec.Ports[i]
-		if int(p.Port) == port.IntValue() {
-			return sm.insert(svc, p)
-		}
-		if port.String() == p.Name {
-			return sm.insert(svc, p)
-		}
-	}
-	return nil
-}
-
-func (sm *serviceMap) insert(svc *v1.Service, port *v1.ServicePort) *Service {
-	if sm._services == nil {
-		sm._services = make(map[portmeta]*Service)
-	}
-	up := parseUpstreamProtocols(svc.Annotations, annotationUpstreamProtocol, "h2", "h2c")
-	protocol := up[port.Name]
-	if protocol == "" {
-		protocol = up[strconv.Itoa(int(port.Port))]
-	}
-
-	s := &Service{
-		Object:      svc,
-		ServicePort: port,
-		Protocol:    protocol,
-
-		MaxConnections:     parseAnnotation(svc.Annotations, annotationMaxConnections),
-		MaxPendingRequests: parseAnnotation(svc.Annotations, annotationMaxPendingRequests),
-		MaxRequests:        parseAnnotation(svc.Annotations, annotationMaxRequests),
-		MaxRetries:         parseAnnotation(svc.Annotations, annotationMaxRetries),
-	}
-	sm._services[s.toMeta()] = s
-	return s
-}
-
-// compute builds a new *dag.dag.
-func (d *DAG) compute() *dag {
-	sm := serviceMap{
-		services: d.services,
-	}
-	service := sm.lookup
-
-	// memoise access to a secrets map, built
-	// as needed from the list of secrets cached
-	// from k8s.
-	_secrets := make(map[meta]*Secret)
-	secret := func(m meta) *Secret {
-		if s, ok := _secrets[m]; ok {
-			return s
-		}
-		sec, ok := d.secrets[m]
-		if !ok {
-			return nil
-		}
-		s := &Secret{
-			object: sec,
-		}
-		_secrets[s.toMeta()] = s
-		return s
-	}
-
-	type hostport struct {
-		host string
-		port int
-	}
-
-	// memoise the production of vhost entries as needed.
-	_vhosts := make(map[hostport]*VirtualHost)
-	vhost := func(host string, port int) *VirtualHost {
-		hp := hostport{host: host, port: port}
-		vh, ok := _vhosts[hp]
-		if !ok {
-			vh = &VirtualHost{
-				Port:   port,
-				host:   host,
-				routes: make(map[string]*Route),
-			}
-			_vhosts[hp] = vh
-		}
-		return vh
-	}
-
-	_svhosts := make(map[hostport]*SecureVirtualHost)
-	svhost := func(host string, port int) *SecureVirtualHost {
-		hp := hostport{host: host, port: port}
-		svh, ok := _svhosts[hp]
-		if !ok {
-			svh = &SecureVirtualHost{
-				Port:   port,
-				host:   host,
-				routes: make(map[string]*Route),
-			}
-			_svhosts[hp] = svh
-		}
-		return svh
-	}
-
-	// setup secure vhosts if there is a matching secret
-	// we do this first so that the set of active secure vhosts is stable
-	// during the second ingress pass
-	for _, ing := range d.ingresses {
-		for _, tls := range ing.Spec.TLS {
-			m := meta{name: tls.SecretName, namespace: ing.Namespace}
-			if sec := secret(m); sec != nil {
-				for _, host := range tls.Hosts {
-					svhost(host, 443).secret = sec
-					// process annotations
-					switch ing.ObjectMeta.Annotations["contour.heptio.com/tls-minimum-protocol-version"] {
-					case "1.3":
-						svhost(host, 443).MinProtoVersion = auth.TlsParameters_TLSv1_3
-					case "1.2":
-						svhost(host, 443).MinProtoVersion = auth.TlsParameters_TLSv1_2
-					default:
-						// any other value is interpreted as TLS/1.1
-						svhost(host, 443).MinProtoVersion = auth.TlsParameters_TLSv1_1
-					}
-				}
-			}
-		}
-	}
-
-	// deconstruct each ingress into routes and virtualhost entries
-	for _, ing := range d.ingresses {
-		// should we create port 80 routes for this ingress
-		httpAllowed := httpAllowed(ing)
-
-		// compute websocket enabled routes
-		wr := websocketRoutes(ing)
-
-		// compute timeout for any routes on this ingress
-		timeout := parseAnnotationTimeout(ing.Annotations, annotationRequestTimeout)
-
-		if ing.Spec.Backend != nil {
-			// handle the annoying default ingress
-			r := &Route{
-				path:         "/",
-				Object:       ing,
-				HTTPSUpgrade: tlsRequired(ing),
-				Websocket:    wr["/"],
-				Timeout:      timeout,
-			}
-			m := meta{name: ing.Spec.Backend.ServiceName, namespace: ing.Namespace}
-			if s := service(m, ing.Spec.Backend.ServicePort); s != nil {
-				r.addService(s, nil, "", 0)
-			}
-			if httpAllowed {
-				vhost("*", 80).routes[r.path] = r
-			}
-		}
-
-		for _, rule := range ing.Spec.Rules {
-			// handle Spec.Rule declarations
-			host := rule.Host
-			if host == "" {
-				host = "*"
-			}
-			for n := range rule.IngressRuleValue.HTTP.Paths {
-				path := rule.IngressRuleValue.HTTP.Paths[n].Path
-				if path == "" {
-					path = "/"
-				}
-				r := &Route{
-					path:         path,
-					Object:       ing,
-					HTTPSUpgrade: tlsRequired(ing),
-					Websocket:    wr[path],
-					Timeout:      timeout,
-				}
-
-				m := meta{name: rule.IngressRuleValue.HTTP.Paths[n].Backend.ServiceName, namespace: ing.Namespace}
-				if s := service(m, rule.IngressRuleValue.HTTP.Paths[n].Backend.ServicePort); s != nil {
-					r.addService(s, nil, "", s.Weight)
-				}
-				if httpAllowed {
-					vhost(host, 80).routes[r.path] = r
-				}
-				if _, ok := _svhosts[hostport{host: host, port: 443}]; ok && host != "*" {
-					svhost(host, 443).routes[r.path] = r
-				}
-			}
-		}
-	}
-
-	// process ingressroute documents
-	var status []Status
-	orphaned := make(map[meta]bool)
-	for _, ir := range d.ingressroutes {
-		if ir.Spec.VirtualHost == nil {
-			// delegate ingress route. mark as orphaned if we haven't reached it before.
-			if _, ok := orphaned[meta{name: ir.Name, namespace: ir.Namespace}]; !ok {
-				orphaned[meta{name: ir.Name, namespace: ir.Namespace}] = true
-			}
-			continue
-		}
-
-		// ensure root ingressroute lives in allowed namespace
-		if !d.rootAllowed(ir) {
-			status = append(status, Status{Object: ir, Status: "invalid", Description: "root IngressRoute cannot be defined in this namespace"})
-			continue
-		}
-
-		host := ir.Spec.VirtualHost.Fqdn
-		if len(strings.TrimSpace(host)) == 0 {
-			status = append(status, Status{Object: ir, Status: "invalid", Description: "Spec.VirtualHost.Fqdn must be specified"})
-			continue
-		}
-
-		if tls := ir.Spec.VirtualHost.TLS; tls != nil {
-			// attach secrets to TLS enabled vhosts
-			m := meta{name: tls.SecretName, namespace: ir.Namespace}
-			if sec := secret(m); sec != nil {
-				svhost(host, 443).secret = sec
-				svhost(host, 443).MinProtoVersion = auth.TlsParameters_TLSv1_1 // TODO(dfc) issue 467
-			}
-		}
-
-		prefixMatch := ""
-		irp := ingressRouteProcessor{
-			host:          host,
-			service:       service,
-			vhost:         vhost,
-			ingressroutes: d.ingressroutes,
-			orphaned:      orphaned,
-		}
-		sts := irp.process(ir, prefixMatch, nil)
-		status = append(status, sts...)
-	}
-
-	var _d dag
-	for _, vh := range _vhosts {
-		_d.roots = append(_d.roots, vh)
-	}
-	for _, svh := range _svhosts {
-		_d.roots = append(_d.roots, svh)
-	}
-
-	for meta, orph := range orphaned {
-		if orph {
-			ir, ok := d.ingressroutes[meta]
-			if ok {
-				status = append(status, Status{Object: ir, Status: "orphaned", Description: "this IngressRoute is not part of a delegation chain from a root IngressRoute"})
-			}
-		}
-	}
-	_d.statuses = status
-	return &_d
-}
-
-// returns true if the root ingressroute lives in a root namespace
-func (d *DAG) rootAllowed(ir *ingressroutev1.IngressRoute) bool {
-	if len(d.IngressRouteRootNamespaces) == 0 {
-		return true
-	}
-	for _, ns := range d.IngressRouteRootNamespaces {
-		if ns == ir.Namespace {
-			return true
-		}
-	}
-	return false
-}
-
-type ingressRouteProcessor struct {
-	host          string
-	service       func(m meta, port intstr.IntOrString) *Service
-	vhost         func(host string, port int) *VirtualHost
-	ingressroutes map[meta]*ingressroutev1.IngressRoute
-	orphaned      map[meta]bool
-}
-
-func (irp *ingressRouteProcessor) process(ir *ingressroutev1.IngressRoute, prefixMatch string, visited []*ingressroutev1.IngressRoute) []Status {
-	visited = append(visited, ir)
-
-	var status []Status
-	for _, route := range ir.Spec.Routes {
-		// route cannot both delegate and point to services
-		if len(route.Services) > 0 && route.Delegate.Name != "" {
-			return []Status{{Object: ir, Status: "invalid", Description: fmt.Sprintf("route %q: cannot specify services and delegate in the same route", route.Match)}}
-		}
-		// base case: The route points to services, so we add them to the vhost
-		if len(route.Services) > 0 {
-			if !matchesPathPrefix(route.Match, prefixMatch) {
-				return []Status{{Object: ir, Status: "invalid", Description: fmt.Sprintf("the path prefix %q does not match the parent's path prefix %q", route.Match, prefixMatch)}}
-			}
-			r := &Route{
-				path:   route.Match,
-				Object: ir,
-			}
-			for _, s := range route.Services {
-				if s.Port < 1 || s.Port > 65535 {
-					return []Status{{Object: ir, Status: "invalid", Description: fmt.Sprintf("route %q: service %q: port must be in the range 1-65535", route.Match, s.Name)}}
-				}
-				if s.Weight < 0 {
-					return []Status{{Object: ir, Status: "invalid", Description: fmt.Sprintf("route %q: service %q: weight must be greater than or equal to zero", route.Match, s.Name)}}
-				}
-				m := meta{name: s.Name, namespace: ir.Namespace}
-				if svc := irp.service(m, intstr.FromInt(s.Port)); svc != nil {
-					r.addService(svc, s.HealthCheck, s.Strategy, s.Weight)
-				}
-			}
-			irp.vhost(irp.host, 80).routes[r.path] = r
-			continue
-		}
-
-		// otherwise, if the route is delegating to another ingressroute, find it and process it.
-		if route.Delegate.Name != "" {
-			namespace := route.Delegate.Namespace
-			if namespace == "" {
-				// we are delegating to another IngressRoute in the same namespace
-				namespace = ir.Namespace
-			}
-			dest, ok := irp.ingressroutes[meta{name: route.Delegate.Name, namespace: namespace}]
-			if ok {
-				// dest is not an orphaned route, as there is an IR that points to it
-				irp.orphaned[meta{name: dest.Name, namespace: dest.Namespace}] = false
-
-				// ensure we are not following an edge that produces a cycle
-				var path []string
-				for _, vir := range visited {
-					path = append(path, fmt.Sprintf("%s/%s", vir.Namespace, vir.Name))
-				}
-				for _, vir := range visited {
-					if dest.Name == vir.Name && dest.Namespace == vir.Namespace {
-						path = append(path, fmt.Sprintf("%s/%s", dest.Namespace, dest.Name))
-						description := fmt.Sprintf("route creates a delegation cycle: %s", strings.Join(path, " -> "))
-						return []Status{{Object: ir, Status: "invalid", Description: description}}
-					}
-				}
-
-				// follow the link and process the target ingress route
-				status = append(status, irp.process(dest, route.Match, visited)...)
-			}
-		}
-	}
-	return append(status, Status{Object: ir, Status: "valid", Description: "valid IngressRoute"})
-}
-
-// matchesPathPrefix checks whether the given path matches the given prefix
-func matchesPathPrefix(path, prefix string) bool {
-	if len(prefix) == 0 {
-		return true
-	}
-	// an empty string cannot have a prefix
-	if len(path) == 0 {
-		return false
-	}
-	if prefix[len(prefix)-1] != '/' {
-		prefix = prefix + "/"
-	}
-	if path[len(path)-1] != '/' {
-		path = path + "/"
-	}
-	return strings.HasPrefix(path, prefix)
 }
 
 type Route struct {
@@ -616,8 +135,12 @@ func (s *SecureVirtualHost) Visit(f func(Vertex)) {
 	f(s.secret)
 }
 
-type Vertex interface {
+type Visitable interface {
 	Visit(func(Vertex))
+}
+
+type Vertex interface {
+	Visitable
 }
 
 // Service represents a K8s Sevice as a DAG vertex. A Service is
@@ -691,11 +214,4 @@ func (s *Secret) toMeta() meta {
 		name:      s.object.Name,
 		namespace: s.object.Namespace,
 	}
-}
-
-// Status contains the status for an IngressRoute (valid / invalid / orphan, etc)
-type Status struct {
-	Object      *ingressroutev1.IngressRoute
-	Status      string
-	Description string
 }
